Add support for testing yarn cluster. Extracted JobManager's and TaskManager's testing messages into stackable traits.
