+++ flink-yarn-tests/src/main/java/org/apache/flink/yarn/TestingApplicationMaster.java
@@ -32,4 +32,10 @@ public class TestingApplicationMaster extends ApplicationMasterBase {
 	public Class<? extends MemoryArchivist> getArchivistClass() {
 		return TestingMemoryArchivist.class;
 	}
+
+	public static void main(String[] args) {
+		TestingApplicationMaster applicationMaster = new TestingApplicationMaster();
+
+		applicationMaster.run(args);
+	}
 }
