+++ flink-yarn/src/main/scala/org/apache/flink/yarn/ApplicationMaster.scala
@@ -20,7 +20,7 @@ package org.apache.flink.yarn
 import org.apache.flink.runtime.jobmanager.{MemoryArchivist, JobManager}
 
 class ApplicationMaster extends ApplicationMasterBase {
-  override def getJobManagerClass: Class[_ <: JobManager] = classOf[JobManager]
+  override def getJobManagerClass: Class[_ <: JobManager] = classOf[YarnJobManager]
   override def getArchivistClass: Class[_ <: MemoryArchivist] = classOf[MemoryArchivist]
 }
